+++ cli/src/main/java/org/jboss/as/cli/handlers/jms/JmsTopicAddHandler.java
@@ -94,8 +94,10 @@ public class JmsTopicAddHandler extends BatchModeCommandHandler {
         }
 
         final String name = this.name.getValue(args, true);
+        String serverName = "default"; // TODO read server name from props
 
         builder.addNode("subsystem", "messaging");
+        builder.addNode("hornetq-server", serverName);
         builder.addNode("jms-topic", name);
         builder.setOperationName("add");
