+++ cli/src/main/java/org/jboss/as/cli/handlers/jms/JmsTopicRemoveHandler.java
@@ -108,8 +108,10 @@ public class JmsTopicRemoveHandler extends BatchModeCommandHandler {
         }
 
         final String name = this.name.getValue(ctx.getParsedCommandLine(), true);
+        String serverName = "default"; // TODO read server name from props
 
         builder.addNode("subsystem", "messaging");
+        builder.addNode("hornetq-server", serverName);
         builder.addNode("jms-topic", name);
         builder.setOperationName("remove");
