+++ demos/legacy/src/main/java/org/jboss/as/demos/client/jms/runner/ExampleRunner.java
@@ -67,6 +67,7 @@ public class ExampleRunner {
             ModelNode op = new ModelNode();
             op.get("operation").set("add");
             op.get("address").add("subsystem", "messaging");
+            op.get("address").add("hornetq-server", "default");
             op.get("address").add("jms-queue", QUEUE_NAME);
             op.get("entries").add(QUEUE_NAME);
             applyUpdate(op, client);
@@ -125,6 +126,7 @@ public class ExampleRunner {
                 ModelNode op = new ModelNode();
                 op.get("operation").set("remove");
                 op.get("address").add("subsystem", "messaging");
+                op.get("address").add("hornetq-server", "default");
                 op.get("address").add("jms-queue", QUEUE_NAME);
                 applyUpdate(op, client);
             }
