+++ demos/legacy/src/main/java/org/jboss/as/demos/client/messaging/runner/ExampleRunner.java
@@ -71,6 +71,7 @@ public class ExampleRunner {
             ModelNode op = new ModelNode();
             op.get("operation").set("add");
             op.get("address").add("subsystem", "messaging");
+            op.get("address").add("hornetq-server", "default");
             op.get("address").add("queue", queueName);
             op.get("queue-address").set(queueName);
             applyUpdate(op, client);
@@ -106,6 +107,7 @@ public class ExampleRunner {
             op = new ModelNode();
             op.get("operation").set("remove");
             op.get("address").add("subsystem", "messaging");
+            op.get("address").add("hornetq-server", "default");
             op.get("address").add("queue", queueName);
             applyUpdate(op, client);
