+++ messaging/src/main/java/org/jboss/as/messaging/MessagingSubsystemRemove.java
@@ -22,22 +22,23 @@
 
 package org.jboss.as.messaging;
 
-import org.hornetq.core.server.group.impl.GroupBinding;
-import org.jboss.as.controller.AbstractRemoveStepHandler;
 import org.jboss.as.controller.OperationContext;
 import org.jboss.as.controller.OperationFailedException;
 import org.jboss.as.controller.OperationStepHandler;
 import org.jboss.as.controller.PathAddress;
+import org.jboss.as.controller.PathElement;
 import org.jboss.as.controller.descriptions.DescriptionProvider;
 import org.jboss.as.controller.registry.Resource;
-import org.jboss.as.messaging.jms.JMSService;
-import org.jboss.as.messaging.jms.JMSServices;
 import org.jboss.dmr.ModelNode;
 
 import java.util.Locale;
+import java.util.Set;
 
 /**
+ * Removes the messaging subsystem.
+ *
  * @author Emanuel Muckenhuber
+ * @author Brian Stansberry (c) 2011 Red Hat Inc.
  */
 class MessagingSubsystemRemove implements OperationStepHandler, DescriptionProvider {
 
@@ -51,42 +52,23 @@ class MessagingSubsystemRemove implements OperationStepHandler, DescriptionProvi
             @Override
             public void execute(final OperationContext context, final ModelNode operation) throws OperationFailedException {
 
-                for(final Resource.ResourceEntry jmsQueue : resource.getChildren(CommonAttributes.JMS_QUEUE)) {
-                    context.removeService(JMSServices.JMS_QUEUE_BASE.append(jmsQueue.getName()));
-                }
-                for(final Resource.ResourceEntry jmsTopic : resource.getChildren(CommonAttributes.JMS_TOPIC)) {
-                    context.removeService(JMSServices.JMS_TOPIC_BASE.append(jmsTopic.getName()));
-                }
-                for(final Resource.ResourceEntry cf : resource.getChildren(CommonAttributes.CONNECTION_FACTORY)) {
-                    context.removeService(JMSServices.JMS_CF_BASE.append(cf.getName()));
-                }
-                for(final Resource.ResourceEntry pcf : resource.getChildren(CommonAttributes.POOLED_CONNECTION_FACTORY)) {
-                    context.removeService(MessagingServices.POOLED_CONNECTION_FACTORY_BASE.append(pcf.getName()));
-                }
-                for(final Resource.ResourceEntry queue : resource.getChildren(CommonAttributes.QUEUE)) {
-                    context.removeService(MessagingServices.CORE_QUEUE_BASE.append(queue.getName()));
+                // TODO should we make the runtime change by default, or require a header indicating that's valid?
+
+                Set<String> serverNames = resource.getChildrenNames(CommonAttributes.HORNETQ_SERVER);
+                for (String serverName : serverNames) {
+                    PathElement path = PathElement.pathElement(CommonAttributes.HORNETQ_SERVER, serverName);
+                    Resource server = resource.getChild(path);
+                    HornetQServerRemove.removeHornetQServer(serverName, context, server);
                 }
 
-                context.removeService(JMSServices.JMS_MANAGER);
-                context.removeService(MessagingServices.JBOSS_MESSAGING);
-                for(final Resource.ResourceEntry broadcastGroup : resource.getChildren(CommonAttributes.BROADCAST_GROUP)) {
-                    context.removeService(GroupBindingService.BROADCAST.append(broadcastGroup.getName()));
-                }
-                for(final Resource.ResourceEntry divertGroup : resource.getChildren(CommonAttributes.DISCOVERY_GROUP)) {
-                    context.removeService(GroupBindingService.DISCOVERY.append(divertGroup.getName()));
-                }
-                context.removeService(MessagingSubsystemAdd.PATH_BASE.append(MessagingSubsystemAdd.DEFAULT_BINDINGS_DIR));
-                context.removeService(MessagingSubsystemAdd.PATH_BASE.append(MessagingSubsystemAdd.DEFAULT_JOURNAL_DIR));
-                context.removeService(MessagingSubsystemAdd.PATH_BASE.append(MessagingSubsystemAdd.DEFAULT_LARGE_MESSSAGE_DIR));
-                context.removeService(MessagingSubsystemAdd.PATH_BASE.append(MessagingSubsystemAdd.DEFAULT_PAGING_DIR));
 
-                context.completeStep();
+                if(context.completeStep() == OperationContext.ResultAction.ROLLBACK) {
+                    //  TODO recover
+                }
             }
         }, OperationContext.Stage.RUNTIME);
 
-        if(context.completeStep() == OperationContext.ResultAction.ROLLBACK) {
-            //
-        }
+        context.completeStep();
     }
 
     @Override
