+++ testsuite/integration/src/test/java/org/jboss/as/testsuite/integration/messaging/mgmt/AddressControlManagementTestCase.java
@@ -185,12 +185,14 @@ public class AddressControlManagementTestCase {
     private ModelNode getSubsystemOperation(String operationName) {
         final ModelNode address = new ModelNode();
         address.add("subsystem", "messaging");
+        address.add("hornetq-server", "default");
         return org.jboss.as.controller.operations.common.Util.getEmptyOperation(operationName, address);
     }
 
     private ModelNode getAddressOperation(String operationName) {
         final ModelNode address = new ModelNode();
         address.add("subsystem", "messaging");
+        address.add("hornetq-server", "default");
         address.add("core-address", getAddress());
         return org.jboss.as.controller.operations.common.Util.getEmptyOperation(operationName, address);
     }
