+++ testsuite/integration/src/test/java/org/jboss/as/testsuite/integration/messaging/mgmt/JMSQueueManagementTestCase.java
@@ -359,6 +359,7 @@ public class JMSQueueManagementTestCase {
     private ModelNode getQueueOperation(String operationName) {
         final ModelNode address = new ModelNode();
         address.add("subsystem", "messaging");
+        address.add("hornetq-server", "default");
         address.add("jms-queue", getQueueName());
         return org.jboss.as.controller.operations.common.Util.getEmptyOperation(operationName, address);
     }
