+++ testsuite/integration/src/test/java/org/jboss/as/testsuite/integration/messaging/mgmt/JMSTopicManagementTestCase.java
@@ -296,6 +296,7 @@ public class JMSTopicManagementTestCase {
     private ModelNode getTopicOperation(String operationName) {
         final ModelNode address = new ModelNode();
         address.add("subsystem", "messaging");
+        address.add("hornetq-server", "default");
         address.add("jms-topic", getTopicName());
         return org.jboss.as.controller.operations.common.Util.getEmptyOperation(operationName, address);
     }
