+++ testsuite/smoke/src/test/java/org/jboss/as/test/embedded/demos/client/jms/JmsClientTestCase.java
@@ -90,6 +90,7 @@ public class JmsClientTestCase {
             ModelNode op = new ModelNode();
             op.get("operation").set("add");
             op.get("address").add("subsystem", "messaging");
+            op.get("address").add("server", "default");
             op.get("address").add("jms-queue", QUEUE_NAME);
             op.get("entries").add(QUEUE_NAME);
             applyUpdate(op, client);
@@ -153,6 +154,7 @@ public class JmsClientTestCase {
                 ModelNode op = new ModelNode();
                 op.get("operation").set("remove");
                 op.get("address").add("subsystem", "messaging");
+                op.get("address").add("server", "default");
                 op.get("address").add("jms-queue", QUEUE_NAME);
                 applyUpdate(op, client);
             }
